+++ example/lib/pages/demo/cascader.dart
@@ -13,7 +13,7 @@ class CascaderPage extends StatefulWidget {
 
   CascaderPage({Key? key}) : super(key: key);
 
-  final String title = "Cascader";
+  final String title = "级联选择器";
 
   @override
   _CascaderPageState createState() => _CascaderPageState();
@@ -22,12 +22,17 @@ class CascaderPage extends StatefulWidget {
 class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin{
 
   //计算拼接地址字符串的
-  List<int> _pickerSelected = [0,0,0];
-  String  _address3 = "请选择";
+  String  _chooseText = "请选择";
+  String  _chooseAddress = "请选择地址";
+
+  List<List<Map>> _textData = [[{"label": "不满意", "value": "0"},{"label": "满意", "value": "1"},{"label": "非常满意", "value": "2"}]];  //数据
+  List<int> _textIndex = [0];            //选择的索引
 
   //{"label": "请选择", "value": ""},
   List<List<Map>> _regionData = [[],[],[]];    //数据
   List<String> _regionValue = ["0","0","0"];   //默认值
+  List<int> _regionIndex = [0,0,0];            //选择的索引
+
 
   @override
   void initState() {
@@ -41,7 +46,7 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
         isRouteFlutter = arguments["flutter"];
         if(!Config.isNative || isRouteFlutter){
           //加载数据
-          _loadData(0,0);
+          _initData(0,0);
         }
       }
     });
@@ -49,72 +54,51 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
   }
 
 
-  ///get请求数据
-  Future<void> _loadData(int pid,int index) async {
+  ///初始化数据
+  Future<void> _initData(int pid,int index) async {
 
-    List<Map> listNew = [
-    ];
+    List<Map> listNew = [];
+
     for (var i = 0; i < 20; i++) {
       listNew.add({"label": pid.toString() + "-" + i.toString(), "value": i.toString()});
       //判断默认值
       if(i.toString() == _regionValue[index]){
-        _pickerSelected[index] = i;
+        _regionIndex[index] = i;
       }
     }
-    if(index == 0){
-      setState(() {
-        _regionData[0] = listNew;
-      });
-      print("update list1...............");
 
-      //接着查询市
-      _loadData(int.parse(_regionValue[0]),1);
+    setState(() {
+      _regionData[index] = listNew;
+    });
 
-    }else if(index == 1){
-      print("update list2...............");
-      setState(() {
-        _regionData[1] = listNew;
-      });
-      //接着查询区
-      _loadData(int.parse(_regionValue[1]),2);
+    _regionValue[index] = _regionData[index][0]['value'];
 
-    }else if(index == 2){
-      print("update list3...............");
-      setState(() {
-        _regionData[2] = listNew;
-      });
+    if(index == 0 || index == 1){
+      //接着查询下级
+      _initData(int.parse(_regionValue[index]),index + 1);
     }
 
   }
 
+  ///切换数据
   Future<void> _changeData(Function dialogState,int pid,int index) async {
-    List<Map> listNew = [
-    ];
+
+    List<Map> listNew = [];
 
     for (var i = 0; i < 20; i++) {
       listNew.add({"label": pid.toString() + "-" + i.toString(), "value": i.toString()});
     }
 
+    dialogState(() {
+      _regionData[index] = listNew;
+      _regionIndex[index] = _regionIndex[index] > _regionData[index].length-1?_regionData[index].length-1:_regionIndex[index];
+    });
+
+    _regionValue[index] = _regionData[index][0]['value'];
+
     if(index == 1){
-      print("update list2...............");
-      dialogState(() {
-        _regionData[1] = listNew;
-      });
-
-      //父节点变化后，选中第一个
-      _regionValue[1] = _regionData[1][0]['value'];
-      //接着查询区
-      _loadData(pid,2);
-
-    }else if(index == 2){
-      print("update list3...............");
-      dialogState(() {
-        _regionData[2] = listNew;
-      });
-
-      //父节点变化后，选中第一个
-      _regionValue[2] = _regionData[2][0]['value'];
-
+      //接着查询下级
+      _changeData(dialogState,int.parse(_regionValue[index]),index + 1);
     }
   }
 
@@ -144,33 +128,22 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
               top: appBarHeight,
               left:0,
               width:  MediaQuery.of(context).size.width,
-              child:Padding(
-                padding: EdgeInsets.only(top: 10, left: 16, right: 16, bottom: 16),
-                child:Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  crossAxisAlignment: CrossAxisAlignment.center,
-                  children: [
-                    Container(
-                      width: 100,
-                      child: Padding(
-                          padding: EdgeInsets.only(top:0,left: 16,bottom: 0),
-                          child:Text(
-                            "级联组件",
-                            textAlign:TextAlign.left,
-                            style: const TextStyle(
-                              color: Colors.black54,
-                              fontSize: 15,
-                            ),
-                          )
-                      ),
-                    ),
-                    Expanded(
-                        flex: 1,
-                        child:GestureDetector(
+              height:  MediaQuery.of(context).size.height - appBarHeight,
+              child:Column(
+                children: [
+
+                  Padding(
+                    padding: EdgeInsets.only(top: 10, left: 16, right: 16, bottom: 16),
+                    child:Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        Container(
+                          width: 100,
                           child: Padding(
-                              padding: EdgeInsets.only(top:0,left:16,right: 16,bottom: 0),
+                              padding: EdgeInsets.only(top:0,left: 16,bottom: 0),
                               child:Text(
-                                _address3,
+                                "一级选择",
                                 textAlign:TextAlign.left,
                                 style: const TextStyle(
                                   color: Colors.black54,
@@ -178,12 +151,73 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
                                 ),
                               )
                           ),
-                          onTap:_openPickerDialog ,
+                        ),
+                        Expanded(
+                            flex: 1,
+                            child:GestureDetector(
+                              child: Padding(
+                                  padding: EdgeInsets.only(top:0,left:16,right: 16,bottom: 0),
+                                  child:Text(
+                                    _chooseText,
+                                    textAlign:TextAlign.left,
+                                    style: const TextStyle(
+                                      color: Colors.black54,
+                                      fontSize: 15,
+                                    ),
+                                  )
+                              ),
+                              onTap:_openTextPickerDialog ,
+                            )
                         )
-                    )
-                  ],
-                ),
+                      ],
+                    ),
+                  ),
+
+                  Padding(
+                    padding: EdgeInsets.only(top: 10, left: 16, right: 16, bottom: 16),
+                    child:Row(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      children: [
+                        Container(
+                          width: 100,
+                          child: Padding(
+                              padding: EdgeInsets.only(top:0,left: 16,bottom: 0),
+                              child:Text(
+                                "三级选择",
+                                textAlign:TextAlign.left,
+                                style: const TextStyle(
+                                  color: Colors.black54,
+                                  fontSize: 15,
+                                ),
+                              )
+                          ),
+                        ),
+                        Expanded(
+                            flex: 1,
+                            child:GestureDetector(
+                              child: Padding(
+                                  padding: EdgeInsets.only(top:0,left:16,right: 16,bottom: 0),
+                                  child:Text(
+                                    _chooseAddress,
+                                    textAlign:TextAlign.left,
+                                    style: const TextStyle(
+                                      color: Colors.black54,
+                                      fontSize: 15,
+                                    ),
+                                  )
+                              ),
+                              onTap:_openAddressPickerDialog ,
+                            )
+                        )
+                      ],
+                    ),
+                  ),
+
+                ],
               ),
+
+
             ),
           ],
         ),
@@ -192,35 +226,53 @@ class _CascaderPageState extends State<CascaderPage> with SingleNativeStateMixin
   }
 
   // 对话框式底部滑动窗口
-  Future _openPickerDialog() async {
+  Future _openTextPickerDialog() async {
     YmUiUtil.showBottomDialog(context,StatefulBuilder(
       builder: (context,state){
         return  YmCascader(
-            _regionData,_regionValue,
+          _textData,
+          currentIndex: [1],
+          onOkClick:(){
+            Navigator.of(context).pop();
+            setState(() {
+              _chooseText = _textData[0].elementAt(_textIndex[0])['label'];
+            });
+          },
+          onCancelClick:(){
+            Navigator.of(context).pop();
+          },
+          onChanged: (position,index,value){
+            _textIndex[position] = index;
+          },
+        );
+      },
+    ));
+  }
+
+  // 对话框式底部滑动窗口
+  Future _openAddressPickerDialog() async {
+    YmUiUtil.showBottomDialog(context,StatefulBuilder(
+      builder: (context,state){
+        return  YmCascader(
+            _regionData,
+            currentIndex: _regionIndex,
             onOkClick:(){
               Navigator.of(context).pop();
               setState(() {
-                _address3 = _regionData[0].elementAt(_pickerSelected[0])['label'] + "/" + _regionData[1].elementAt(_pickerSelected[1])['label'] + "/" + _regionData[2].elementAt(_pickerSelected[2])['label'];
+                _chooseAddress = _regionData[0].elementAt(_regionIndex[0])['label'] + "/" + _regionData[1].elementAt(_regionIndex[1])['label'] + "/" + _regionData[2].elementAt(_regionIndex[2])['label'];
               });
             },
             onCancelClick:(){
               Navigator.of(context).pop();
             },
             onChanged: (position,index,value){
-              if(position == 0){
-                _pickerSelected[0] = index;
-                _pickerSelected[1] = 0;
-                _pickerSelected[2] = 0;
-                _regionValue[0] = value;
-                _changeData(state,int.parse(value),1);
-              }else if(position == 1){
-                _pickerSelected[1] = index;
-                _pickerSelected[2] = 0;
-                _regionValue[1] = value;
-                _changeData(state,int.parse(value),2);
-              }else if(position == 2){
-                _pickerSelected[2] = index;
-                _regionValue[2] = value;
+
+              //保存选中的值
+              _regionIndex[position] = index;
+              _regionValue[position] = value;
+
+              if(position == 0 || position == 1){
+                _changeData(state,int.parse(value),position + 1);
               }
             },
         );